requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
asgiref==3.11.1
sqlparse==0.5.5
psycopg2-binary==2.9.11
pillow==11.0.0
PyJWT==2.11.0
pytz==2024.2
tzdata==2024.2
django-cors-headers